create cluster object in all streaming metrics test cases
