+++ test/distributed/org/apache/cassandra/distributed/test/metrics/StreamingMetricsRepairTest.java
@@ -18,6 +18,27 @@
 
 package org.apache.cassandra.distributed.test.metrics;
 
-public class StreamingMetricsRepairTest
+import org.junit.Test;
+
+import org.apache.cassandra.distributed.Cluster;
+import org.apache.cassandra.distributed.test.TestBaseImpl;
+
+import static org.apache.cassandra.distributed.api.Feature.NETWORK;
+import static org.apache.cassandra.distributed.test.metrics.StreamingMetricsTestUtils.testMetricsWithStreamingFromTwoNodes;
+
+public class StreamingMetricsRepairTest extends TestBaseImpl
 {
+    @Test
+    public void testMetricsWithRepairAndStreamingFromTwoNodes() throws Exception
+    {
+        try(Cluster cluster = init(Cluster.build(3)
+                                          .withDataDirCount(1)
+                                          .withConfig(config -> config.with(NETWORK)
+                                                                      .set("stream_entire_sstables", false)
+                                                                      .set("hinted_handoff_enabled", false))
+                                          .start(), 2))
+        {
+            testMetricsWithStreamingFromTwoNodes(true, cluster);
+        }
+    }
 }
